+++ src/apps/sequencer/engine/generators/RandomGenerator.h
@@ -20,7 +20,7 @@ public:
         uint16_t seed = 0;
         uint8_t smooth = 0;
         int8_t bias = 0;
-        uint8_t scale = 10;
+        uint8_t scale = 5;
     };
 
     RandomGenerator(SequenceBuilder &builder, Params &params);