+++ src/apps/sequencer/model/CurveTrack.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "BaseTrack.h"
+#include "BaseTrackPatternFollow.h"
 #include "Config.h"
 #include "Types.h"
 #include "CurveSequence.h"
@@ -8,13 +10,11 @@
 #include "FileDefs.h"
 #include "core/utils/StringUtils.h"
 
-class CurveTrack {
+class CurveTrack : public BaseTrack, public BaseTrackPatternFollow {
 public:
     //----------------------------------------
     // Types
     //----------------------------------------
-
-    static constexpr size_t NameLength = FileHeader::NameLength; 
 
     typedef std::array<CurveSequence, CONFIG_PATTERN_COUNT + CONFIG_SNAPSHOT_COUNT> CurveSequenceArray;
 
@@ -61,13 +61,6 @@ public:
     //----------------------------------------
     // Properties
     //----------------------------------------
-
-    // name
-        // name
-    const char *name() const { return _name; }
-    void setName(const char *name) {
-        StringUtils::copy(_name, name, sizeof(_name));
-    }
 
     // playMode
 
@@ -241,7 +234,6 @@ private:
     }
 
     int8_t _trackIndex = -1;
-    char _name[NameLength + 1];
     Types::PlayMode _playMode;
     FillMode _fillMode;
     MuteMode _muteMode;