+++ src/apps/sequencer/ui/pages/NoteSequencePage.cpp
@@ -1,5 +1,6 @@
 #include "NoteSequencePage.h"
 
+#include "ListPage.h"
 #include "Pages.h"
 
 #include "ui/LedPainter.h"
@@ -65,6 +66,12 @@ void NoteSequencePage::keyPress(KeyPressEvent &event) {
 
     if (!event.consumed()) {
         ListPage::keyPress(event);
+    }
+    if (key.isEncoder()) {
+        auto row = ListPage::selectedRow();
+        if (row == 5) {
+            _listModel.setSelectedScale(_project.scale());
+        }
     }
 }