+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-## v 0.1.46 (4 January 2024)
+# v0.1.4.47 (24 January 2024)
+- launchpad circuit mode improvements
+- random generator: random seed just on init method
+- patter follow
+- pattern chain quick shortcut from pattern page
+- scale edit: scales are changed only if the encoder is pressed
+- on scale change the sequence notes are changed according to the previous scale. if a note in the previous scale is also present in the new scale the value is preeserved. if it is not present the nearest note in the new scale will be selected
+- launchpad follows pattern in song mode
+
+## v0.1.46 (4 January 2024)
 
 - UI note edit page reaggment
 - song sync with clock